`sql<number>` yang tak dicast itu STRING; dan gerbang CI merah 7 jam tiap hari
Dua temuan, keduanya diukur langsung terhadap Postgres 16 yang dipakai repo
ini — bukan diperkirakan dari membaca kode.

── 1. `sql<number>` adalah janji yang tak diperiksa siapa pun ──────────────

Driver `pg` memulangkan `numeric` sebagai STRING, dan bukan cuma `numeric`:

    SUM(numeric) → "3000.75" string      MAX(integer) → 7   number
    SUM(integer) → "10"      string      MIN(integer) → 3   number
    AVG(integer) → "5.0000…" string      COUNT(*)     → "2" string
    MAX(numeric) → "2.50"    string      GREATEST(int,int) → 7 number

Tak ada aturan yang bisa dipegang seseorang di kepalanya: hasilnya bergantung
pada tipe balikan tiap agregat atas tipe tiap kolom. `MAX` aman, `SUM` tidak,
dan keduanya ditulis bersebelahan.

Yang membuatnya berbahaya: TYPESCRIPT BERPIHAK PADA YANG SALAH. `sql<number>`
membuat nilainya bertipe `number` sementara isinya string, jadi kompilernya
justru menjamin sesuatu yang tak benar. Terukur: `a + b` atas dua nilai yang
sama memulangkan "3000.753000.75". Dan nilai yang lolos ke `c.json()`
melanggar kontrak API — klien menerima string di tempat `LaporanHarian`
menjanjikan number; di mobile `as num` MELEMPAR, `as num?` memulangkan null.

Dua belas ekspresi tanpa cast, semuanya di jalur uang atau cacah. Repo ini
selamat hari ini karena DISIPLIN — tiap pemanggil membungkusnya `Number(...)`,
dan kuperiksa keduabelasnya satu per satu: semua benar. Yang kuubah bukan
kebenarannya melainkan DI MANA ia tinggal: cast di ekspresinya sendiri,
sekali, di tempat yang tak bisa dilupakan pemanggil berikutnya.

Cast-nya TIDAK mengubah satu angka pun, dan itu diukur bukan diasumsikan:
`x::float8` menghasilkan float64 yang BIT-PER-BIT sama dengan `Number("x")` —
diuji atas 99.999.999,99 dan 0,01, seluruh barisnya `Object.is`-identik.

Satu pengecualian, dengan alasannya: `GREATEST(0, EXTRACT(EPOCH …))` di
laporan durasi tak dicast. Ia tak pernah dipilih mentah — hanya disisipkan ke
ekspresi yang sudah ber-`::int` — dan mencastnya ke float8 BUKAN netral seperti
yang lain: `ROUND` atas numeric membulatkan setengah menjauhi nol, atas float8
ke genap terdekat.

Penjaganya `test/sql-number-bukan-janji.test.ts`. Bukti merah: satu cast
dicabut → menuduh tepat berkas & barisnya; pengecualiannya diubah → menuduh
daftarnya basi.

Penjaga itu sempat MENUDUH TULISANKU SENDIRI: prosa di `porsi-ditagih.ts` yang
menjelaskan aturan ini memuat `sql<number>` sebagai contoh. Penjaga yang
menuduh dokumentasi lebih buruk daripada tak ada penjaga — ia membuat menulis
penjelasan yang baik jadi hal yang memerahkan CI. Komentar kini dibuang lebih
dulu, dengan pasangan yang membuktikan pembuangnya tak ikut memakan kode.

── 2. verify-api.sh merah tujuh jam setiap hari ────────────────────────────

§218 memakai `date -u -d '6 days ago' +%F`, sementara server menghitung
tunggakan pada tanggal bisnis (Asia/Jakarta). Antara 17.00–24.00 UTC — saat
Jakarta sudah berganti hari — "6 hari lalu" versi UTC berjarak TUJUH hari dari
hari ini versi WIB.

Bukan hipotesis: gerbangnya gagal saat kujalankan pukul 23.18 UTC, tunggakan
8×3=24 alih-alih 7×3=21, "sisa 76 harusnya 79". Sesudah diperbaiki dan
dijalankan lagi di jam yang sama: 2.688 lolos, 0 gagal, §218 memulangkan 79.

Bentuk yang sama dengan temuan-temuan sesi ini: ATURANNYA SUDAH TERTULIS DAN
BERKOMENTAR. Delapan belas dari dua puluh perhitungan tanggal-saja sudah
memakai `TZ=Asia/Jakarta`, dua di antaranya bahkan menuliskan alasannya
("kontainer bisa UTC", "tanggal bisnis, bukan tanggal UTC server"). Dua pintu
terlewat — §218 dan §80.

Yang paling mahal bukan kegagalannya melainkan bentuknya: gerbang yang merah
tujuh jam sehari tanpa sebab nyata mengajari orang mengabaikan warna merah.

Penjaganya menumpang `zona-waktu-satu-suara.test.ts` — konsepnya memang satu.
Bukti merahnya menemukan LUBANG DI PENJAGAKU SENDIRI: versi pertama memeriksa
per BARIS, dan `DARI80=…; SAMPAI80=…` menaruh dua perhitungan di satu baris,
jadi satu `TZ=` di paruh kedua meloloskan paruh pertama yang telanjang. Kini
dipecah per perintah; kedua pintu terbukti tertuduh.

── Verifikasi ──────────────────────────────────────────────────────────────

  · npm run typecheck — bersih
  · npm test — 1.996 lolos, 159 berkas
  · scripts/verify-api.sh terhadap Postgres SEGAR — 2.688 lolos, 0 gagal.
    Ini yang benar-benar membuktikan cast-nya: SQL baru tak terlihat
    typecheck, dan satu salah ketik `::float8` hanya muncul saat dijalankan.

+++ apps/server/src/modules/admin-error-log/routes.ts
@@ -57,7 +57,7 @@ async function ambilKelompok(sejak: Date, saring?: "4xx" | "5xx", cari?: string)
       terakhir_pada: sql<string>`max(${errorLogs.waktu})`,
       jumlah_user: sql<number>`count(distinct ${errorLogs.userId})::int`,
       jumlah_perusahaan: sql<number>`count(distinct ${errorLogs.companyId})::int`,
-      status: sql<number>`(array_agg(${errorLogs.status} ORDER BY ${errorLogs.waktu} DESC))[1]`,
+      status: sql<number>`(array_agg(${errorLogs.status} ORDER BY ${errorLogs.waktu} DESC))[1]::int`,
       metode: sql<string>`(array_agg(${errorLogs.metode} ORDER BY ${errorLogs.waktu} DESC))[1]`,
       jalur_pola: sql<string>`(array_agg(${errorLogs.jalurPola} ORDER BY ${errorLogs.waktu} DESC))[1]`,
       pesan: sql<string>`(array_agg(${errorLogs.pesan} ORDER BY ${errorLogs.waktu} DESC))[1]`,
